Handle paths that don't exist in `erase`

+++ collegram/channels.py
@@ -564,8 +564,11 @@ def erase(channel_paths: ChannelPaths, fs: AbstractFileSystem = LOCAL_FS,):
     Remove all data about channel from disk (for deleted channels or those who became
     private).
     '''
-    fs.rm(channel_paths.anon_map)
-    fs.rm(channel_paths.channel)
-    for p in fs.ls(channel_paths.messages):
-        fs.rm(p)
-    fs.rmdir(channel_paths.messages)
+    if fs.exists(channel_paths.anon_map):
+        fs.rm(channel_paths.anon_map)
+    if fs.exists(channel_paths.channel):
+        fs.rm(channel_paths.channel)
+    if fs.exists(channel_paths.messages):
+        for p in fs.ls(channel_paths.messages):
+            fs.rm(p)
+        fs.rmdir(channel_paths.messages)
